👌 Code review : OK - Revisando codigo de errores

+++ src/build/routers/film.ts
@@ -8,25 +8,31 @@ const routerFilm = express.Router();
 
 routerFilm.get("/:id", async (req:Request, res:Response)=>{
   let filmId:number = parseInt(req.params.id)
-  await refillPeopleForThisFilm(res,filmId)
-  let filmsRepository = await AppDataSource.getRepository(Films)
-  let film = await filmsRepository.findOneBy({id: filmId})
-  if (film === null) {
-    saveError(404,"No se encontró la película, (Not found Error).")
-    res.render("infoTemplate",{error: error})
-  }else{
-    let peopleIds = await getPeopleIdWhitFilmId(film.id)
-    let peopleRepository = await AppDataSource.getRepository(People)
-    let characters = await peopleRepository 
-      .createQueryBuilder("film")
-      .where("id IN (:...ids)", { ids: peopleIds })
-      .getMany();
-    if(characters.length > 0){
-      res.render("infoTemplate",{film: film, characters: characters})
+  if(!isNaN(filmId)){
+    await refillPeopleForThisFilm(res,filmId)
+    let filmsRepository = await AppDataSource.getRepository(Films)
+    let film = await filmsRepository.findOneBy({id: filmId})
+    if (film === null) {
+      saveError(res, 404, `No se encontró la película ${req.params.id}, (Not found Error).`)
+      res.render("infoTemplate",{error: error})
     }else{
-      res.render("infoTemplate",{film: film, error: error})
+      let peopleIds = await getPeopleIdWhitFilmId(film.id)
+      let peopleRepository = await AppDataSource.getRepository(People)
+      let characters = await peopleRepository 
+        .createQueryBuilder("film")
+        .where("id IN (:...ids)", { ids: peopleIds })
+        .getMany();
+      if(characters.length > 0){
+        res.render("infoTemplate",{film: film, characters: characters})
+      }else{
+        res.render("infoTemplate",{film: film, error: error})
+      }
     }
+  }else{
+    saveError(res, 400,`La solicitud "${req.params.id}" es incorrecta.`)
+    res.render("infoTemplate",{error: error})
   }
+
 })
 
 routerFilm.get("/s/search", async (req:Request, res:Response) => {
@@ -43,7 +49,7 @@ routerFilm.get("/s/search", async (req:Request, res:Response) => {
   if(films.length > 0){
     res.render("infoTemplate",{results: films})
   }else{
-    saveError(404, "Not Found")
+    saveError(res, 404, "Not Found")
     res.render("infoTemplate",{error: error})
   }
 })
@@ -60,31 +66,36 @@ routerFilm.get("/s/all", async (req:Request, res:Response) => {
 })
 
 routerFilm.delete("/del/:id", async (req:Request, res:Response)=>{
-  let paramNumber = parseInt(req.params.id);
-  let filmsRepository = await AppDataSource.getRepository(Films)
-  let film = await filmsRepository.findOneBy({id: paramNumber})
-  if(film){
-    let charactersIdsToDelete = await getPeopleIdWhitFilmId(film.id);
-    let peopleRepository = await AppDataSource.getRepository(People)
-    await peopleRepository.createQueryBuilder()
-      .delete()
-      .from(People)
-      .where("id IN (:...charactersIdsToDelete)", { charactersIdsToDelete })
-      .execute();
-    let filmId = [film.id]
-    let peopleInFilmsRepository = await AppDataSource.getRepository(PeopleInFilms)
-    await peopleInFilmsRepository.createQueryBuilder()
-      .delete()
-      .from(PeopleInFilms)
-      .where("film_id IN (:...filmId)", { filmId })
-      .execute();
-    await updateFilmCharactersStatus(film.id,false)
+  let filmId = parseInt(req.params.id);
+  if(!isNaN(filmId)){
+    let filmsRepository = await AppDataSource.getRepository(Films)
+    let film = await filmsRepository.findOneBy({id: filmId})
+    if(film){
+      let charactersIdsToDelete = await getPeopleIdWhitFilmId(film.id);
+      let peopleRepository = await AppDataSource.getRepository(People)
+      await peopleRepository.createQueryBuilder()
+        .delete()
+        .from(People)
+        .where("id IN (:...charactersIdsToDelete)", { charactersIdsToDelete })
+        .execute();
+      let filmId = [film.id]
+      let peopleInFilmsRepository = await AppDataSource.getRepository(PeopleInFilms)
+      await peopleInFilmsRepository.createQueryBuilder()
+        .delete()
+        .from(PeopleInFilms)
+        .where("film_id IN (:...filmId)", { filmId })
+        .execute();
+      await updateFilmCharactersStatus(film.id,false)
+      let films = await filmsRepository.find()
+      res.render("infoTemplate", {results: films});
+    }else{
+      saveError(res, 404, `La película ${req.params.id} a eliminar no se encuentra (Not Found)`);
+      res.render("infoTemplate", {error: error});
+    }
   }else{
-    saveError(404,'La película buscada no se encuentra (Not Found)');
-    res.render("infoTemplate", {error: error});
+    saveError(res, 400, `La solicitud "${req.params.id}" es incorrecta.`)
+    res.render("infoTemplate",{error: error})
   }
-  let films = await filmsRepository.find()
-  res.render("infoTemplate", {results: films});
 })
 
 routerFilm.delete("/s/del/all",async(req:Request,res:Response) =>{
